Willbranch

+++ CAS_WEB/cas/views.py
@@ -204,19 +204,17 @@ def search_project_by_id(request, id):
 
 @login_required
 def update_project(request, id):
-    content = dict()
-    content['user'] = request.user
-    try:
-        project = Project.objects.filter(id=id).first()
-    except:
-        raise Http404("Project not found")
+    if request.method == 'GET':
+        content = dict()
+        content['user'] = request.user
+        try:
+            project = Project.objects.filter(id=id).first()
+        except:
+            raise Http404("Project not found")
 
-    content['project'] = project
-    return render(request, 'cas/update_project.html', content)
+        content['project'] = project
+        return render(request, 'cas/update_project.html', content)
 
-
-@login_required
-def update_project_info(request, id):
     project = get_object_or_404(Project, pk=id)
     name = request.POST['name']
     description = request.POST['description']
@@ -234,15 +232,12 @@ def update_project_info(request, id):
 
 @login_required
 def delete_project(request, id):
-    project = get_object_or_404(Project, pk=id)
-    content = dict()
-    content['user'] = request.user
-    content['project'] = project
-    return render(request, 'cas/delete_project.html', content)
-
-
-@login_required
-def __delete_project(request, id):
+    if request.method == 'GET':
+        project = get_object_or_404(Project, pk=id)
+        content = {}
+        content['user'] = request.user
+        content['project'] = project
+        return render(request, 'cas/delete_project.html', content)
     project = get_object_or_404(Project, pk=id)
     project.delete()
